Moved a bunch of data processing structs out of ProcessDataLogger and into their own sub-package.
