+++ src/flightmill/ProcessDataLoggerFile.java
@@ -36,6 +36,10 @@ import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
+import flightmill.dataStorageStructs.InputCommandLine;
+import flightmill.dataStorageStructs.InputDataLine;
+import flightmill.dataStorageStructs.IntermediateDataLine;
+
 /**
  * This programs takes a csv file from a datalogger attached to a flight mill
  * and collates it.
@@ -467,179 +471,4 @@ public class ProcessDataLoggerFile {
 
         return newFile.getAbsolutePath();
     }//end unzipFile(zipFileName)
-
-    // class for input data directly mapped to input file
-    protected static class InputDataLine {
-
-        double time = 0.0;
-        double channels[] = null;
-
-        public InputDataLine(String inputLine) {
-            String items[] = inputLine.split(",");
-            time = Double.parseDouble(items[0]);
-            channels = new double[items.length - 1];
-            for (int idx = 1; idx < items.length; idx++) {
-                channels[idx - 1] = Double.parseDouble(items[idx]);
-            }
-        }//end 1-arg constructor
-
-        public InputDataLine(double time, int channel, int numChannel) {
-            this.time = time;
-            channels = new double[numChannel];
-            channels[channel] = 1.0;
-        }//end 3-arg constructor
-    }//end class InputDataLine
-
-    // class to hold data lines that hold only peaks
-    protected static class IntermediateDataLine {
-
-        int channel;
-        double elapsedTime;
-        double value;
-        int peakWidth = 1;
-
-        public IntermediateDataLine(int channel, double elapsedTime, double value) {
-            this.channel = channel;
-            this.elapsedTime = elapsedTime;
-            this.value = value;
-        }//end 3-arg constructor
-    }//end class IntermediateDataLine
-
-    // class to hold parsed input command options
-    protected static class InputCommandLine {
-
-        private String inputFileName = "ns-1-26-23-meeting-test.csv";
-        private String outputFileName = "ns-1-26-23-meeting-test.csv.out";
-        private int skipLines = 3;    // header lines at top of data logger file
-        private double timeInterval = 0.01;  // time between data logger readings
-        private int numberOfChannelsUsed = 0;    // self explanitory
-        private boolean zipFileFlg = false;              // zip input file
-        private boolean dataTimeFlg = true;  // add date time to each output line
-        private boolean peakWidthFlg = false; // add date time to each output line
-        private double threshold = 1.5;
-
-        //<editor-fold defaultstate="collapsed" desc="getters/setters">
-        /**
-         * @return the inputFileName
-         */
-        public String getInputFileName() {
-            return inputFileName;
-        }
-
-        /**
-         * @param aInputFileName the inputFileName to set
-         */
-        public void setInputFileName(String aInputFileName) {
-            inputFileName = aInputFileName;
-        }
-
-        /**
-         * @return the outputFileName
-         */
-        public String getOutputFileName() {
-            return outputFileName;
-        }
-
-        /**
-         * @param aOutputFileName the outputFileName to set
-         */
-        public void setOutputFileName(String aOutputFileName) {
-            outputFileName = aOutputFileName;
-        }
-
-        /**
-         * @return the skipLines
-         */
-        public int getSkipLines() {
-            return skipLines;
-        }
-
-        /**
-         * @param aSkipLines the skipLines to set
-         */
-        public void setSkipLines(int aSkipLines) {
-            skipLines = aSkipLines;
-        }
-
-        /**
-         * @return the timeInterval
-         */
-        public double getTimeInterval() {
-            return timeInterval;
-        }
-
-        /**
-         * @param aTimeInterval the timeInterval to set
-         */
-        public void setTimeInterval(double aTimeInterval) {
-            timeInterval = aTimeInterval;
-        }
-
-        /**
-         * @return the numberOfChannelsUsed
-         */
-        public int getNumberOfChannelsUsed() {
-            return numberOfChannelsUsed;
-        }
-
-        /**
-         * @param aNumberOfChannelsUsed the numberOfChannelsUsed to set
-         */
-        public void setNumberOfChannelsUsed(int aNumberOfChannelsUsed) {
-            numberOfChannelsUsed = aNumberOfChannelsUsed;
-        }
-
-        /**
-         * @return the zipFileFlg
-         */
-        public boolean isZipFileFlg() {
-            return zipFileFlg;
-        }
-
-        /**
-         * @param aZipFileFlg the zipFileFlg to set
-         */
-        public void setZipFileFlg(boolean aZipFileFlg) {
-            zipFileFlg = aZipFileFlg;
-        }
-
-        /**
-         * @return the dataTimeFlg
-         */
-        public boolean isDataTimeFlg() {
-            return dataTimeFlg;
-        }
-
-        /**
-         * @param dataTimeFlg the dataTimeFlg to set
-         */
-        public void setDataTimeFlg(boolean dataTimeFlg) {
-            this.dataTimeFlg = dataTimeFlg;
-        }
-
-        /**
-         * @return the peakWidthFlg
-         */
-        public boolean isPeakWidthFlg() {
-            return peakWidthFlg;
-        }
-
-        /**
-         * @param peakWidthFlg the peakWidthFlg to set
-         */
-        public void setPeakWidthFlg(boolean peakWidthFlg) {
-            this.peakWidthFlg = peakWidthFlg;
-        }
-
-        public double getThreshold() {
-            return threshold;
-        }
-
-        public void setThreshold(double threshold) {
-            this.threshold = threshold;
-        }
-
-//</editor-fold>
-    }//end class InputCommandLine
-
 }//end class ProcessDataLoggerFile
